github-notifications/requirements.txt
functions-framework==3.*
flask==2.1.0
slackclient==2.9.4
python-dotenv==0.21.0